% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompBiomass.R
\name{CompBiomass}
\alias{CompBiomass}
\title{Compare Biomass of two model runs of the same model}
\usage{
CompBiomass(file1, file2, groups)
}
\arguments{
\item{x}{Description of parameter x}
}
\value{
Description of the returned object
CompBiomass('E:/Atlantis/output/Atlantis_2025_03/M672025_03_2025_04_22_01/OutBiomIndx.txt',
'E:/Atlantis/output/Atlantis_2025_03/M672025_03_2025_03_27_01/OutBiomIndx.txt',
'E:/Atlantis/output/Atlantis_2025_03/M672025_03_2025_03_27_01/GroupsIceland.csv')
my_function(10)
}
\description{
#'
}
