% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompBiomassN.R
\name{CompBiomassN}
\alias{CompBiomassN}
\title{Compares biomass for fish, mammals, sharks, and birds from >1 different Atlantis runs for one species}
\usage{
CompBiomassN(files, groups, yrstart)
}
\arguments{
\item{x}{Description of parameter x}
}
\value{
Description of the returned object
}
\description{
Compares biomass for fish, mammals, sharks, and birds from >1 different Atlantis runs for one species
}
\examples{
CompBiomassN(c('E:/Atlantis/output/Atlantis_2025_03/M672025_03_2025_04_30_01/OutBiomIndx.txt',
'E:/Atlantis/output/Atlantis_2025_03/M672025_03_2025_04_30_22/OutBiomIndx.txt'),
'E:/Atlantis/output/Atlantis_2025_03/M672025_03_2025_04_30_01/GroupsIceland.csv',
2020)
}
