% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompareCatch.R
\name{CompCatch}
\alias{CompCatch}
\title{Compares catch from different Atlantis runs}
\usage{
CompCatch(file1, file2, groupvector)
}
\arguments{
\item{x}{Description of parameter x}
}
\value{
Description of the returned object
}
\description{
Compares catch from different Atlantis runs
}
\examples{
CompCatch('E:/Atlantis/output/Atlantis6702/M6702_2025_02_10_02/OutCatch.txt',
'E:/Atlantis/output/Atlantis6702/M6702_2025_02_10_03/OutCatch.txt',
c('FCD', 'FHA'))
}
