% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompareCatchN.R
\name{CompCatchN}
\alias{CompCatchN}
\title{Compares total catch from >1 different Atlantis runs for all harvested species}
\usage{
CompCatchN(files, groups, styear)
}
\arguments{
\item{x}{Description of parameter x}
}
\value{
Description of the returned object
}
\description{
Compares total catch from >1 different Atlantis runs for all harvested species
}
\examples{
CompCatchN(c('E:/Atlantis/output/Atlantis_2025_03/M672025_03_2025_04_30_01/OutCatchPerFishery.txt',
'E:/Atlantis/output/Atlantis_2025_03/M672025_03_2025_04_30_22/OutCatchPerFishery.txt'),
'E:/Atlantis/output/Atlantis_2025_03/M672025_03_2025_04_30_01/GroupsIceland.csv',
2020)
}
