% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompareCatchPrN.R
\name{CompCatchPRN}
\alias{CompCatchPRN}
\title{Compares pr change  from >1 different Atlantis runs for all harvested species to the status quo management
must indicate sq management projection using whichSQ}
\usage{
CompCatchPRN(files, styear, whichSQ)
}
\arguments{
\item{x}{Description of parameter x}
}
\value{
Description of the returned object
}
\description{
Compares pr change  from >1 different Atlantis runs for all harvested species to the status quo management
must indicate sq management projection using whichSQ
}
\examples{
CompCatchPRN(c('E:/Atlantis/output/Atlantis_2025_03/M672025_03_2025_04_30_01/OutCatchPerFishery.txt',
'E:/Atlantis/output/Atlantis_2025_03/M672025_03_2025_04_30_22/OutCatchPerFishery.txt'),
2020, 'M672025_03_2025_04_30_01')
}
