% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompareCatchPerFishery.R
\name{CompCPF}
\alias{CompCPF}
\title{Compares catch per fishery from >1 different Atlantis runs for one species}
\usage{
CompCPF(files, sp, styear)
}
\arguments{
\item{x}{Description of parameter x}
}
\value{
Description of the returned object
}
\description{
Compares catch per fishery from >1 different Atlantis runs for one species
}
\examples{
CompCPF(c('E:/Atlantis/output/Atlantis_2025_03/M672025_03_2025_04_30_01/OutCatchPerFishery.txt',
       'E:/Atlantis/output/Atlantis_2025_03/M672025_03_2025_04_30_22/OutCatchPerFishery.txt'),
"FCD",
2020)
}
