% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompareGroupsCsv.R
\name{CompGroupFiles}
\alias{CompGroupFiles}
\title{Compare Groups csvs}
\usage{
CompGroupFiles(file1, file2, file1group, file2group)
}
\arguments{
\item{x}{Description of parameter x}
}
\value{
Description of the returned object
CompMigFiles("E:/Atlantis/input/atlantis6702/GroupsIceland6702.csv",
"E:/Atlantis/NoBa_params/nordic_groups.csv",
"WMW",  "MWH")
my_function(10)
}
\description{
compares groups csvs from different Atlantis models
}
