% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompareMigCsv.R
\name{CompMigFiles}
\alias{CompMigFiles}
\title{Compare Migration csvs}
\usage{
CompMigFiles(file1, file2, file1group, file2group)
}
\arguments{
\item{x}{Description of parameter x}
}
\value{
Description of the returned object
CompMigFiles("E:/Atlantis/input/atlantis6702/migration_6702_2025_02_07_01.csv",
"E:/Atlantis/NoBa_params/NoBa_migrations.csv",
"WMW",  "MWH")
my_function(10)
}
\description{
compares migration csvs from different Atlantis models
}
