% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CatchPF.R
\name{CPF}
\alias{CPF}
\title{Compares catch from different Atlantis runs}
\usage{
CPF(file1, file2, groups)
}
\arguments{
\item{x}{Description of parameter x}
}
\value{
Description of the returned object
}
\description{
Compares catch from different Atlantis runs
}
\examples{
CompCatch('E:/Atlantis/output/Atlantis_2025_03/M672025_03_2025_04_22_01/OutCatch.txt',
'E:/Atlantis/output/Atlantis_2025_03/M672025_03_2025_03_27_01/OutCatch.txt',
'E:/Atlantis/output/Atlantis_2025_03/M672025_03_2025_03_27_01/GroupsIceland.csv')
}
