% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_parameters.R
\name{extract_parameters}
\alias{extract_parameters}
\title{extract_parameters}
\usage{
extract_parameters(file_path, key)
}
\arguments{
\item{x}{Description of parameter x}
}
\value{
Description of the returned object
}
\description{
extracts parameters of interest from biological files
}
\examples{
my_function(10)
}
