% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractNums.R
\name{extractNums}
\alias{extractNums}
\title{extract NAA from two different model runs}
\usage{
extractNums(outdir1, outdir2, toutinc, startyear)
}
\arguments{
\item{x}{Description of parameter x}
}
\value{
Description of the returned object
extractNums('E:/Atlantis/output/AtlantisFeb25/M67F25_2025_03_03_01/',
'E:/Atlantis/output/Atlantis6610/M6610_0212_01/', 365, 1948)
my_function(10)
}
\description{
extract NAA from two different model runs
}
