% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MigSpawn.R
\name{MigSpawnRecTime}
\alias{MigSpawnRecTime}
\title{Calculates Migration, spawning and recruitment time from biological.prm and migration.csv}
\usage{
MigSpawnRecTime(bfile, speciescode, mfile, matstage)
}
\arguments{
\item{x}{Description of parameter x}
}
\value{
Description of the returned object
}
\description{
Calculates Migration, spawning and recruitment time from biological.prm and migration.csv
}
\examples{
MigSpawnRecTime("E:/Atlantis/input/atlantis6702/B391ES3JK6.6702_noWH.prm", "WMW",
"E:/Atlantis/input/atlantis6702/migration_6702_2025_02_10_01.csv", 1)
}
