% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YOY.R
\name{yoytable}
\alias{yoytable}
\title{reads yoy file for species of choice}
\usage{
yoytable(file, groupvector)
}
\arguments{
\item{x}{Description of parameter x}
}
\value{
Description of the returned object
}
\description{
reads yoy file for species of choice
}
\examples{
yoytable("E:/Atlantis/output/Atlantis6702/M6702_2025_02_10_01/OutYOY.txt", c('FMI', 'WMW', 'WHB'))
}
